✨ feature : 페이지 라우팅, 리코일 적용
home페이지에서 검색하면 search 페이지로 이동해서 검색결과 렌더링
