+++ package.json
@@ -10,7 +10,10 @@
     "node-fetch": "^3.3.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.20.1",
     "react-scripts": "5.0.1",
+    "recoil": "^0.7.7",
+    "styled-components": "^6.1.1",
     "web-vitals": "^2.1.4"
   },
   "proxy": "https://openapi.naver.com",
