+++ src/Pages/Home/home.styled.jsx
@@ -1,0 +1,3 @@
+import { styled } from 'styled-components';
+
+export default HomeWrapper = styled.div``;
