+++ src/Pages/Programming/ProgrammingPage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 
-export default function Programming() {
+export default function ProgrammingPage() {
   return <div>Programming</div>;
 }
