--- src/Pages/Search/Search.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function Search() {
-  return <div>Search</div>;
-}
