+++ src/Recoil/searchResultState.jsx
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const searchResultState = atom({
+  key: 'searchResultState',
+  default: null,
+});
