+++ src/index.js
@@ -2,7 +2,12 @@ import React from 'react';
 // import ReactDOM from 'react-dom'; //구버전
 import { createRoot } from 'react-dom/client';
 import App from './App';
+import { BrowserRouter } from 'react-router-dom';
 
 const container = document.getElementById('root');
 const root = createRoot(container);
-root.render(<App />);
+root.render(
+  <BrowserRouter>
+    <App />
+  </BrowserRouter>,
+);
